+++ pkg/client/daemon/outbound_linux.go
@@ -77,6 +77,7 @@ func (o *outbound) runOverridingServer(c context.Context, onReady func()) error 
 	}
 
 	o.setSearchPathFunc = func(c context.Context, paths []string) {
+		paths = append(paths, "svc.cluster.local.", "cluster.local.", "")
 		o.search = paths
 	}